requirements.txt
Scrapy
python-dotenv
sqlalchemy
psycopg2-binary